fix(e2e): port GB300 scenario to current harness API
The e2e harness moved from RunScenario(t, &Scenario{}) to declarative
Register(&Scenario{}); Validator now returns error and Scenario has no .T.
Rewrite the GB300 managed-driver scenario accordingly (compile break after
rebase onto main: undefined RunScenario, Validator signature, s.T).

Gate it behind RUN_GB300_E2E=true (SkipIf) so the shared CI e2e run skips the
18-node GB300 rack it has no quota/capacity for; a DRI runs it against a
GB300-capable subscription. Renamed _test.go -> .go to register at package init.
